Add robot launch file
The robot.launch file will start a ros enviorenment and also currently
launches the teleop nodes.  This will likely be the main launch file for
the robot.
    
The file currently supports the argument robot_addr which can be used by
adding robot_addr:=127.0.0.1 (example) when you call the launch file. It
currently defaults to localhost.
    
Closes #17