Add model files
Add the model file for the ground of the IGVC.  It has ms paint quality
lines spray painted on a high res tiled grass picture.  It also acts as
the ground plane.  For gazebo to use the model it must be copied into
the .gazebo/models folder.  The script install_models.sh should do this
depending on how you have your enviorenment setup.