Add localization package
The localization package will contain code to help find the robots
position.  It currently houses the orientation node which is empty and
simply spins forever while ros is running.

A custom message has also been added which represents the euler angle
orientation of the robot and has three components: yaw, pitch and roll.